separate activity and BoundService
