+++ maui/Platforms/Android/IPosSystemTransport.cs
@@ -1,0 +1,9 @@
+﻿using fiskaltrust.Middleware.Demo.Models;
+
+namespace fiskaltrust.Middleware.Demo.Platforms.Android
+{
+    public interface IPosSystemTransport
+    {
+        public Task<PosSystemApiResponse> SendAsync(RequestInfo requestInfo);
+    }
+}
