+++ maui/Platforms/Android/PosSystemTransportFactory.cs
@@ -1,0 +1,17 @@
+﻿namespace fiskaltrust.Middleware.Demo.Platforms.Android
+{
+    public static class PosSystemTransportFactory
+    {
+        public static IPosSystemTransport GetInstance(bool useBoundService)
+        {
+            if (useBoundService)
+            {
+                return new BoundServiceTransport();
+            }
+            else {
+                return new ActivityTransport();
+            }
+
+        }
+    }
+}
